+++ app/Livewire/ForgottenPassword.php
@@ -3,10 +3,12 @@
 namespace App\Livewire;
 
 use Illuminate\Support\Facades\Password;
+use Livewire\Attributes\Layout;
 use Livewire\Attributes\Title;
 use Livewire\Attributes\Validate;
 use Livewire\Component;
 
+#[Layout('components.layouts.app')]
 #[Title('Reset your password')]
 class ForgottenPassword extends Component
 {